♻️refactor SentenceAccumulator
made it more ergonomic, and returned sentences from the addToken and completeSentence functions when sentences are available. This is better than forcing it to take a flume channel, as it's more useful for other programs, and I, as the function caller, can just pass it to a flume channel after it gives me a result, which is my application-specific.