🐛reduce race condition occurence
recv_future_ordered() is not cancel safe, because while futures_ordered.next() appears to be cancel safe. If recv_future_ordered() is canceled between getting the result from futures_ordered.next() and returning it, the value will be lost.

Removing the sleep statement between the two operations to make them almost instantaneous has made it so that I am not perceiving any loss of values. but the race condition still exists and can still occur.