🐛fix crash on playing failure sound
Tokio spawn was crashing, saying that a Tokio runtime had not been started yet. This, I believe, is because it is being spawned in a static lazy lock context, and did not inside the main function, which the program's main Tokio reactor is associated with. Spawning a normal thread seems to fix this.